refactor registration logic into utilities file
