+++ QuantumGravPy/src/QuantumGrav/gnnblock.py
@@ -1,111 +1,9 @@
 # torch
 import torch
 import torch_geometric.nn as tgnn
-
+from . import utils
 # quality of life
 from typing import Any
-
-gnn_layers: dict[str, torch.nn.Module] = {
-    "gcn": tgnn.conv.GCNConv,
-    "gat": tgnn.conv.GATConv,
-    "sage": tgnn.conv.SAGEConv,
-    "graph": tgnn.conv.GraphConv,
-}
-
-normalizer_layers: dict[str, torch.nn.Module] = {
-    "identity": torch.nn.Identity,
-    "batch_norm": torch.nn.BatchNorm1d,
-    "layer_norm": torch.nn.LayerNorm,
-}
-
-activation_layers: dict[str, torch.nn.Module] = {
-    "relu": torch.nn.ReLU,
-    "leaky_relu": torch.nn.LeakyReLU,
-    "sigmoid": torch.nn.Sigmoid,
-    "tanh": torch.nn.Tanh,
-    "identity": torch.nn.Identity,
-}
-
-
-def register_gnn_layer(gnn_layer_name: str, gnn_layer: torch.nn.Module) -> None:
-    """Register a GNN layer with the module
-
-    Args:
-        gnn_layer_name (str): The name of the GNN layer.
-        gnn_layer (torch.nn.Module): The GNN layer to register.
-    """
-    if gnn_layer_name in gnn_layers:
-        raise ValueError(f"GNN layer '{gnn_layer_name}' is already registered.")
-    gnn_layers[gnn_layer_name] = gnn_layer
-
-
-def register_normalizer(
-    normalizer_name: str, normalizer_layer: torch.nn.Module
-) -> None:
-    """Register a normalizer layer with the module
-
-    Args:
-        normalizer_name (str): The name of the normalizer.
-        normalizer_layer (torch.nn.Module): The normalizer layer to register.
-
-    Raises:
-        ValueError: If the normalizer layer is already registered.
-    """
-    if normalizer_name in normalizer_layers:
-        raise ValueError(f"Normalizer '{normalizer_name}' is already registered.")
-    normalizer_layers[normalizer_name] = normalizer_layer
-
-
-def register_activation(
-    activation_name: str, activation_layer: torch.nn.Module
-) -> None:
-    """Register an activation layer with the module
-
-    Args:
-        activation_name (str): The name of the activation layer.
-        activation_layer (torch.nn.Module): The activation layer to register.
-
-    Raises:
-        ValueError: If the activation layer is already registered.
-    """
-    if activation_name in activation_layers:
-        raise ValueError(f"Activation '{activation_name}' is already registered.")
-    activation_layers[activation_name] = activation_layer
-
-
-def get_registered_gnn_layer(name: str) -> torch.nn.Module | None:
-    """Get a registered GNN layer by name.
-    Args:
-        name (str): The name of the GNN layer.
-
-    Returns:
-        torch.nn.Module | None: The registered GNN layer named `name`, or None if not found.
-    """
-    return gnn_layers[name] if name in gnn_layers else None
-
-
-def get_registered_normalizer(name: str) -> torch.nn.Module | None:
-    """Get a registered normalizer layer by name.
-
-    Args:
-        name (str): The name of the normalizer layer.
-
-    Returns:
-        torch.nn.Module | None: The registered normalizer layer named `name`, or None if not found.
-    """
-    return normalizer_layers[name] if name in normalizer_layers else None
-
-
-def get_registered_activation(name: str) -> torch.nn.Module | None:
-    """Get a registered activation layer by name.
-
-    Args:
-        name (str): The name of the activation layer.
-
-    Returns:
-        torch.nn.Module | None: The registered activation layer named `name`, or None if not found.
-    """
-    return activation_layers[name] if name in activation_layers else None
 
 
 class GNNBlock(torch.nn.Module):
@@ -221,9 +119,9 @@ class GNNBlock(torch.nn.Module):
             in_channels=config["in_channels"],
             out_channels=config["out_channels"],
             dropout=config.get("dropout", 0.3),
-            gnn_layer_type=gnn_layers[config["gnn_layer_type"]],
-            normalizer=normalizer_layers[config["normalizer"]],
-            activation=activation_layers[config["activation"]],
+            gnn_layer_type=utils.gnn_layers[config["gnn_layer_type"]],
+            normalizer=utils.normalizer_layers[config["normalizer"]],
+            activation=utils.activation_layers[config["activation"]],
             gnn_layer_args=config.get("gnn_layer_args", []),
             gnn_layer_kwargs=config.get("gnn_layer_kwargs", {}),
             norm_args=config.get("norm_args", []),
